refactor code into firestore helper
